+++ packages/aws-cdk/lib/api/aws-auth/sdk.ts
@@ -321,7 +321,7 @@ import { traceMemberMethods } from './tracing';
 import { defaultCliUserAgent } from './user-agent';
 import { debug } from '../../logging';
 import { AuthenticationError } from '../../toolkit/error';
-import { formatErrorMessage } from '../../util/error';
+import { formatErrorMessage } from '../../util/format-error';
 
 export interface S3ClientOptions {
   /**
