+++ packages/aws-cdk/lib/api/logs/index.ts
@@ -1,0 +1,2 @@
+export * from './logs-monitor';
+export * from './find-cloudwatch-logs';
