+++ packages/aws-cdk/lib/api/resource-import/migrator.ts
@@ -4,9 +4,9 @@ import * as fs from 'fs-extra';
 import { ImportDeploymentOptions, ResourceImporter } from './importer';
 import { info } from '../../cli/messages';
 import type { IIoHost, ToolkitAction } from '../../toolkit/cli-io-host';
+import { formatTime } from '../../util/string-manipulation';
 import { StackCollection } from '../cxapp/cloud-assembly';
 import type { Deployments, ResourcesToImport } from '../deployments';
-import { formatTime } from '../util/string-manipulation';
 
 export interface ResourceMigratorProps {
   deployments: Deployments;
