+++ packages/aws-cdk/lib/api/work-graph/index.ts
@@ -1,0 +1,3 @@
+export * from './work-graph';
+export * from './work-graph-builder';
+export * from './work-graph-types';
