+++ packages/aws-cdk/lib/init.ts
@@ -7,7 +7,7 @@ import { invokeBuiltinHooks } from './init-hooks';
 import { error, info, warning } from './logging';
 import { ToolkitError } from './toolkit/error';
 import { cdkHomeDir, rootDir } from './util/directories';
-import { formatErrorMessage } from './util/error';
+import { formatErrorMessage } from './util/format-error';
 import { rangeFromSemver } from './util/version-range';
 
 /* eslint-disable @typescript-eslint/no-var-requires */ // Packages don't have @types module
