+++ packages/aws-cdk/lib/util/archive.ts
@@ -2,7 +2,7 @@ import { error } from 'console';
 import { createWriteStream, promises as fs } from 'fs';
 import * as path from 'path';
 import * as glob from 'glob';
-import { formatErrorMessage } from './error';
+import { formatErrorMessage } from './format-error';
 
 // eslint-disable-next-line @typescript-eslint/no-require-imports
 const archiver = require('archiver');
