+++ packages/aws-cdk/lib/util/format-error.ts
