+++ packages/aws-cdk/lib/util/index.ts
@@ -1,5 +1,15 @@
+export * from './archive';
 export * from './arrays';
 export * from './bool';
+export * from './bytes';
+export * from './cloudformation';
+export * from './content-hash';
+export * from './format-error';
 export * from './objects';
+export * from './parallel';
+export * from './serialize';
+export * from './string-manipulation';
+export * from './type-brands';
 export * from './types';
-export * from './tables';
+export * from './version-range';
+export * from './yaml-cfn';
