+++ packages/aws-cdk/lib/util/string-manipulation.ts
