+++ packages/aws-cdk/test/util/error.test.ts
@@ -1,4 +1,4 @@
-import { formatErrorMessage } from '../../lib/util/error';
+import { formatErrorMessage } from '../../lib/util/format-error';
 
 describe('formatErrorMessage', () => {
   test('should return the formatted message for a regular Error object', () => {
