+++ packages/aws-cdk/test/util/yaml.test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable import/order */
-import { deserializeStructure, toYAML } from '../lib/serialize';
+import { deserializeStructure, toYAML } from '../../lib/util/serialize';
 
 // Preferred quote of the YAML library
 const q = '"';
